moving /healthcheck endpoint from app.js to config/express.js
